requirements.txt
django
djangorestframework
markdown
django-filter
mysql-python
pytz
django-cors-headers
django-rest-swagger